back-end almost finished

+++ back-end/java/Empresa.java
@@ -1,19 +1,21 @@
 package sql;
 
+import java.math.BigInteger;
+
 public class Empresa extends Usuario {
 	
-	private Integer nrCnpj;
+	private BigInteger nrCnpj;
 
-	public Empresa(Integer id, Integer cnpj, String nome, String email, String senha, String cidade) {
+	public Empresa(Integer id, BigInteger cnpj, String nome, String email, String senha, String cidade) {
 		super(id, nome, email, senha, cidade);
 		setNrCnpj(cnpj);
 	}
 
-	public Integer getNrCnpj() {
+	public BigInteger getNrCnpj() {
 		return nrCnpj;
 	}
 
-	public void setNrCnpj(Integer nrCnpj) {
+	public void setNrCnpj(BigInteger nrCnpj) {
 		this.nrCnpj = nrCnpj;
 	}
 }
